[CL-2140] Remove SignUpIn, move directly into SignUpInModal
