--- front/app/containers/Authentication/SignUpIn/SignUpInModal/SignUpIn.tsx
@@ -1,94 +1,0 @@
-import React from 'react';
-
-// hooks
-import useAppConfiguration from 'hooks/useAppConfiguration';
-import { Box } from '@citizenlab/cl2-component-library';
-
-// component
-import SignUp from './SignUp';
-import SignIn from './SignIn';
-
-// styling
-import { ContextShape } from 'events/verificationModal';
-
-// utils
-import { isNilOrError } from 'utils/helperUtils';
-import { openSignUpInModal } from 'events/openSignUpInModal';
-
-export type TSignUpInFlow = 'signup' | 'signin';
-
-export type TSignUpInError = 'general' | 'franceconnect_merging_failed';
-interface ISignUpInError {
-  code: TSignUpInError;
-}
-
-export interface ISignUpInMetaData {
-  flow: TSignUpInFlow;
-  pathname: string;
-  verification?: boolean;
-  verificationContext?: ContextShape;
-  error?: ISignUpInError;
-  isInvitation?: boolean;
-  token?: string;
-  noAutofocus?: boolean;
-  action?: () => void;
-}
-
-interface Props {
-  metaData: ISignUpInMetaData;
-  onSignUpInCompleted: () => void;
-  className?: string;
-  fullScreen?: boolean;
-}
-
-function getNewFlow(flow: TSignUpInFlow) {
-  switch (flow) {
-    case 'signup':
-      return 'signin';
-    case 'signin':
-      return 'signup';
-  }
-}
-
-const SignUpIn = ({
-  metaData,
-  onSignUpInCompleted,
-  className,
-  fullScreen,
-}: Props) => {
-  const appConfiguration = useAppConfiguration();
-
-  const onToggleSelectedMethod = () => {
-    const flow = getNewFlow(metaData.flow);
-    openSignUpInModal({
-      ...metaData,
-      flow,
-    });
-  };
-
-  if (!isNilOrError(appConfiguration)) {
-    return (
-      <Box className={className}>
-        {metaData.flow === 'signup' ? (
-          <SignUp
-            metaData={metaData}
-            onSignUpCompleted={onSignUpInCompleted}
-            onGoToSignIn={onToggleSelectedMethod}
-            fullScreen={fullScreen}
-          />
-        ) : (
-          <SignIn
-            metaData={metaData}
-            onSignInCompleted={onSignUpInCompleted}
-            onGoToSignUp={onToggleSelectedMethod}
-            fullScreen={fullScreen}
-          />
-        )}
-      </Box>
-    );
-  }
-
-  return null;
-};
-
-export default SignUpIn;
